create pool service and clean controller logic
