+++ src/controller/pool.py
@@ -1,20 +1,14 @@
-from enum import Enum
-from ..utils.queries import Queries
-from ..utils.formatting import Formattor
+from services.pool import PoolService, Pools
+from utils.queries import Queries
+from utils.formatting import Formattor
 
-class Pools(Enum):
-    FRAX = 'frax'
-    SUSD = 'susd'
+
 
 def get_pools():
     try:
-        result = []
-        # @TODO: Move to a service
-        for pool in Pools: 
-            result.append(Queries().get_pool(pool._value_))
-
+        result = PoolService().get_all_pools()
     except Exception as e:
-       print(e)
+       print('error on get_pools', e)
        return Formattor().formatted_response(400,{
             'error': "failed to fetch all pools"
         })
